travelbackend/requirements.txt
asgiref==3.10.0
Django==5.0
django-cors-headers==4.3.0
django-extensions==4.1
django-filter==23.5
django-grappelli==4.0.2
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
